+++ database/migrations/2019_03_10_050953_create_salaries_table.php
@@ -16,7 +16,7 @@ class CreateSalariesTable extends Migration
         Schema::create('salaries', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('employee_id');
-            $table->float('salary_amount');
+            $table->bigInteger('salary_amount');
             $table->foreign('employee_id')->references('id')->on('users');
             $table->timestamps();
         });
